requirements.txt
Crawl4AI
python-dotenv
pydantic
ipython
dspy
mlflow
langchain
langchain-groq
langchain-google-genai
groq
pydantic